moved extract_actions/1 to Utils module
