+++ lib/llm_composer/models/open_router.ex
@@ -9,7 +9,6 @@ defmodule LlmComposer.Models.OpenRouter do
   use Tesla
 
   alias LlmComposer.Errors.MissingKeyError
-  alias LlmComposer.FunctionCall
   alias LlmComposer.LlmResponse
   alias LlmComposer.Models.Utils
 
@@ -97,7 +96,7 @@ defmodule LlmComposer.Models.OpenRouter do
       end
     end
 
-    actions = extract_actions(body)
+    actions = Utils.extract_actions(body)
     {:ok, %{response: body, actions: actions}}
   end
 
@@ -107,26 +106,6 @@ defmodule LlmComposer.Models.OpenRouter do
 
   defp handle_response({:error, reason}, _request_opts) do
     {:error, reason}
-  end
-
-  @spec extract_actions(map()) :: nil | []
-  defp extract_actions(%{"choices" => choices}) when is_list(choices) do
-    choices
-    |> Enum.filter(&(&1["finish_reason"] == "tool_calls"))
-    |> Enum.map(&get_action/1)
-  end
-
-  defp extract_actions(_response), do: []
-
-  defp get_action(%{"message" => %{"tool_calls" => calls}}) do
-    Enum.map(calls, fn call ->
-      %FunctionCall{
-        type: "function",
-        id: call["id"],
-        name: call["function"]["name"],
-        arguments: Jason.decode!(call["function"]["arguments"])
-      }
-    end)
   end
 
   defp get_key do
